Added a number class that is backed by a double. Added a nullable int32 editor. Made the Editing/Refleciton/ReflectedMinMaxEditableProperty.cs use the DoubleBackedNumber to provide min, max and Increment.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@1740 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ Engine/Editing/Refleciton/EditableMinMaxAttribute.cs
@@ -23,7 +23,7 @@ namespace Engine.Editing
         /// Constructor.
         /// </summary>
         /// <param name="type">The type of field this variable wraps if it represents something else. Can be used to show more advanced editors.</param>
-        public EditableMinMaxAttribute(float minValue, float maxValue, float increment, String info = "")
+        public EditableMinMaxAttribute(double minValue, double maxValue, double increment, String info = "")
             :base(info)
         {
             this.MinValue = minValue;
@@ -39,10 +39,10 @@ namespace Engine.Editing
             };
         }
 
-        public float MinValue { get; set; }
+        public double MinValue { get; set; }
 
-        public float MaxValue { get; set; }
+        public double MaxValue { get; set; }
 
-        public float Increment { get; set; }
+        public double Increment { get; set; }
     }
 }
